0.8.4-dev4 client codegen ✨

+++ clients/kotlin/src/main/kotlin/com/print-nanny/client/models/Device.kt
@@ -63,7 +63,7 @@ data class Device (
     @Json(name = "cores")
     val cores: kotlin.Int,
     @Json(name = "ram")
-    val ram: kotlin.Int,
+    val ram: kotlin.Long,
     @Json(name = "cpu_flags")
     val cpuFlags: kotlin.String,
     @Json(name = "id")
