+++ clients/kotlin/src/main/kotlin/com/print-nanny/client/models/DeviceIdentity.kt
@@ -70,7 +70,7 @@ data class DeviceIdentity (
     @Json(name = "cores")
     val cores: kotlin.Int,
     @Json(name = "ram")
-    val ram: kotlin.Int,
+    val ram: kotlin.Long,
     @Json(name = "cpu_flags")
     val cpuFlags: kotlin.String,
     @Json(name = "id")
