+++ clients/kotlin/src/main/kotlin/com/print-nanny/client/models/DeviceRequest.kt
@@ -50,7 +50,7 @@ data class DeviceRequest (
     @Json(name = "cores")
     val cores: kotlin.Int,
     @Json(name = "ram")
-    val ram: kotlin.Int,
+    val ram: kotlin.Long,
     @Json(name = "cpu_flags")
     val cpuFlags: kotlin.String
 ) : Serializable {
