+++ clients/rust/src/models/device_identity.rs
@@ -52,7 +52,7 @@ pub struct DeviceIdentity {
     #[serde(rename = "cores")]
     pub cores: i32,
     #[serde(rename = "ram")]
-    pub ram: i32,
+    pub ram: i64,
     #[serde(rename = "cpu_flags")]
     pub cpu_flags: String,
     #[serde(rename = "url", skip_serializing_if = "Option::is_none")]
@@ -72,7 +72,7 @@ pub struct DeviceIdentity {
 }
 
 impl DeviceIdentity {
-    pub fn new(name: String, os_version: String, os: String, kernel_version: String, hardware: String, revision: String, model: String, serial: String, cores: i32, ram: i32, cpu_flags: String) -> DeviceIdentity {
+    pub fn new(name: String, os_version: String, os: String, kernel_version: String, hardware: String, revision: String, model: String, serial: String, cores: i32, ram: i64, cpu_flags: String) -> DeviceIdentity {
         DeviceIdentity {
             id: None,
             deleted: None,
